changed: - moved and renamed UserService to controllers/UserController.java
