+++ src/main/java/org/kainos/groupA/GroupAApplication.java
@@ -5,9 +5,7 @@ import io.dropwizard.setup.Environment;
 import io.federecio.dropwizard.swagger.SwaggerBundle;
 import io.federecio.dropwizard.swagger.SwaggerBundleConfiguration;
 import org.kainos.groupA.controller.JobRoleController;
-import org.kainos.groupA.api.UserService;
-import org.kainos.groupA.dao.JobRoleDao;
-import org.kainos.groupA.utils.DatabaseConnector;
+import org.kainos.groupA.controller.UserController;
 
 public class GroupAApplication extends io.dropwizard.Application<GroupAConfiguration> {
 
@@ -33,7 +31,7 @@ public class GroupAApplication extends io.dropwizard.Application<GroupAConfigura
     @Override
     public void run(final GroupAConfiguration configuration,
                     final Environment environment) {
-        environment.jersey().register(new UserService());
+        environment.jersey().register(new UserController());
         environment.jersey().register(new JobRoleController());
     }
 }
