+++ src/main/java/org/kainos/groupA/controller/UserController.java
@@ -1,4 +1,4 @@
-package org.kainos.groupA.api;
+package org.kainos.groupA.controller;
 
 import io.swagger.annotations.Api;
 import org.kainos.groupA.models.User;
@@ -12,7 +12,7 @@ import java.util.List;
 
 @Path("/api")
 @Api("User API")
-public class UserService {
+public class UserController {
     //Endpoints
     @GET
     @Path("/users")
